chore: restructure io-related classes into a io directory
