+++ src/main/kotlin/racingcar/io/InputValidator.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.io
 
 object InputValidator {
