+++ src/test/kotlin/racingcar/InputValidatorTest.kt
@@ -2,7 +2,7 @@ import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import racingcar.InputValidator
+import racingcar.io.InputValidator
 
 class InputValidatorTest {
